Rewrite welcome app in zig
This will make the coding of the coding of the app take longer, but it is much smaller. The welcome executable post strip was 11M for the go program, and it is 36K post strip in zig. Making it ~292 times smaller in size, for the exact same outcome of a program.